pyproject.toml
[project]
name = "atzar"
version = "0.1.0"
description = "Weekly random Spotify playlist generator"
authors = [
    {name = "Adrià Luz",email = "adrialuzllompart@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
packages = [{ include = "atzar", from = "src" }]
dependencies = [
    "spotipy (>=2.25.1,<3.0.0)",
    "numpy (>=2.3.2,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
black = "^25.1.0"
flake8 = "^7.3.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
